toolkit/mozapps/downloads/tests/chrome/test_multiword_search.xul

<?xml version="1.0"?>
<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/.  -->
<!--
 * Test for bug 419403 that lets the download manager support multiple word
 * search against the download name, source/referrer, date/time, file size,
 * etc.
-->

<window title="Download Manager Test"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="test();">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="utils.js"/>

  <script type="application/javascript">
  <![CDATA[

function test()
{
  var dmui = getDMUI();
  if (!dmui) {
    todo(false, "skip test for toolkit download manager UI");
    return;
  }

  let dm = Cc["@mozilla.org/download-manager;1"].
           getService(Ci.nsIDownloadManager);
  let db = dm.DBConnection;

  // Empty any old downloads
  db.executeSimpleSQL("DELETE FROM moz_downloads");

  // Make a file name for the downloads
  let file = Cc["@mozilla.org/file/directory_service;1"].
             getService(Ci.nsIProperties).get("TmpD", Ci.nsIFile);
  file.append("multiWord");
  let filePath = Cc["@mozilla.org/network/io-service;1"].
                 getService(Ci.nsIIOService).newFileURI(file).spec;

  let stmt = db.createStatement(
    "INSERT INTO moz_downloads (name, target, source, state, endTime, maxBytes) " +
    "VALUES (?1, ?2, ?3, ?4, ?5, ?6)");

  try {
    for each (let site in ["ed.agadak.net", "mozilla.org"]) {
      stmt.bindByIndex(0, "Super Pimped Download");
      stmt.bindByIndex(1, filePath);
      stmt.bindByIndex(2, "http://" + site + "/file");
      stmt.bindByIndex(3, dm.DOWNLOAD_FINISHED);
      stmt.bindByIndex(4, new Date(1985, 7, 2) * 1000);
      stmt.bindByIndex(5, 111222333444);

      // Add it!
      stmt.execute();
    }
  } finally {
    stmt.reset();
    stmt.finalize();
  }

  // Close the UI if necessary
  let wm = Cc["@mozilla.org/appshell/window-mediator;1"].
           getService(Ci.nsIWindowMediator);
  let win = wm.getMostRecentWindow("Download:Manager");
  if (win) win.close();

  let obs = Cc["@mozilla.org/observer-service;1"].
            getService(Ci.nsIObserverService);
  const DLMGR_UI_DONE = "download-manager-ui-done";

  let testPhase = -1;
  let testObs = {
    observe: function(aSubject, aTopic, aData) {
      if (aTopic != DLMGR_UI_DONE)
        return;

      let win = aSubject.QueryInterface(Ci.nsIDOMWindow);
      win.focus();
      let $ = function(aId) win.document.getElementById(aId);
      let downloadView = $("downloadView");
      let searchbox = $("searchbox");

      let search = function(aTerms) {
        searchbox.value = aTerms;
        searchbox.doCommand();
      };

      let testResults = function(aExpected) {
        is(downloadView.itemCount, aExpected,
           "Phase " + testPhase + ": search matched " + aExpected + " download(s)");
      };

      // The list must have built, so figure out what test to do
      switch (++testPhase) {
        case 0:
          // Search for multiple words in any order in all places
          search("download super pimped 104 GB august 2");

          break;
        case 1:
          // Done populating the two items
          testResults(2);

          // Do partial word matches including the site
          search("Agadak.net aug 2 downl 104");

          break;
        case 2:
          // Done populating the one result
          testResults(1);

          // The search term shouldn't be treated like a regular expression,
          // e.g. "D.wnload" shouldn't match "Download".
          search("D.wnload");

          break;
        case 3:
          testResults(0);

          // We're done!
          win.close();
          obs.removeObserver(testObs, DLMGR_UI_DONE);
          SimpleTest.finish();

          break;
      }
    }
  };
  obs.addObserver(testObs, DLMGR_UI_DONE, false);

  // Show the Download Manager UI
  dmui.show();

  SimpleTest.waitForExplicitFinish();
}

  ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display:none;"></div>
    <pre id="test"></pre>
  </body>
</window>
